Get rid of strcasestr.c vendored file
The initial indea was to move strcasestr.c out of src/common,
as it only used by the main program, perhaps changing to the
OpenBSD implementation at the same time instead of falling back
to the Gnulib test.

Then it appeared we probably don't support systems without strcasestr().
I checked ihe following systems support it:
- glibc
- uClibc    https://git.uclibc.org/uClibc/tree/libc/string/strcasestr.c
- musl libc https://git.musl-libc.org/cgit/musl/tree/src/string/strcasestr.c
- Bionic    https://android.googlesource.com/platform/bionic/+log/master/libc/string/strcasestr.c
- Newlib    https://sourceware.org/git/?p=newlib-cygwin.git;a=blob;f=newlib/libc/string/strcasestr.c
- FreeeBSD  https://cgit.freebsd.org/src/tree/lib/libc/string/strcasestr.c
- OpenBSD   https://cvsweb.openbsd.org/src/lib/libc/string/strcasestr.c
- NetBSD    http://cvsweb.netbsd.org/bsdweb.cgi/src/lib/libc/string/strcasestr.c
- Solaris   https://docs.oracle.com/cd/E88353_01/html/E37843/strcasestr-3c.html

It might appear not to be available when _GNU_SOURCE or equivalent macros
have not been defined. Now with AC_USE_SYSTEM_EXTENSIONS, I doubt the
function is missing on any platform, therefore I get rid of the test and
the Gnulib fallback.

Signed-off-by: Dimitri Papadopoulos <3350651-DimitriPapadopoulos@users.noreply.gitlab.com>