Further improvements to arguments of AC_INIT
The first argument is the full package name. Change it to match the
GitLab home page and documentation:
	ocserv → OpenConnect VPN Server

The package tarname differs from the package name: the latter designates
the full package name, while the former is the distribution tarball name.
Because the tarname cannot be inferred from the newly modified full
package name, we have to set it explicitly:
	ocserv

The last argument url should be the home page for the package.

Signed-off-by: Dimitri Papadopoulos <3350651-DimitriPapadopoulos@users.noreply.gitlab.com>